Memory Fix Release
Fixed a memory problem, at the cost of skipping test status indicators in the schema navigator at the db and schema levels.

Also, fixed a bug where RT thinks a parameter-less procedure within a package actually had a parameter.

Also, removed the hyperlink in the error report, that let you jump to the failing scenario group. To keep its functionality would have been very time consuming and very slow to use, due to the reworked schema navigator.

Note this version was actually produced in June 2018, but not committed till now.